+++ applications/noise-gui/noise-widget.cpp
@@ -7,7 +7,7 @@
 #include "../../regen/gl-types/blit-state.h"
 #include "../../regen/scene/state-configurer.h"
 #include "regen/textures/fbo-state.h"
-#include "../../regen/effects/fullscreen-pass.h"
+#include "../../regen/passes/fullscreen-pass.h"
 #include <regen/animation/animation-manager.h>
 #include <QInputDialog>
