+++ regen/compute/CMakeLists.txt
@@ -5,5 +5,5 @@ file(GLOB HEADER_FILES *.h)
 
 install(
     FILES ${HEADER_FILES}
-    DESTINATION ${HEADER_INSTALL_PATH}/math
+    DESTINATION ${HEADER_INSTALL_PATH}/compute
 )
