+++ regen/memory/CMakeLists.txt
@@ -4,5 +4,5 @@ file(GLOB HEADER_FILES *.h)
 
 install(
     FILES ${HEADER_FILES}
-    DESTINATION ${HEADER_INSTALL_PATH}/buffer
+    DESTINATION ${HEADER_INSTALL_PATH}/memory
 )
