+++ regen/passes/bloom-pass.cpp
