+++ regen/passes/bloom-pass.h
