+++ regen/passes/bloom-texture.cpp
