+++ regen/passes/bloom-texture.h
