+++ regen/passes/geometric-picking.cpp
