+++ regen/passes/geometric-picking.h
