+++ regen/passes/motion-blur.cpp
