+++ regen/scene/CMakeLists.txt
@@ -5,5 +5,5 @@ file(GLOB HEADER_FILES *.h)
 
 install(
     FILES ${HEADER_FILES}
-    DESTINATION ${HEADER_INSTALL_PATH}/scenes
+    DESTINATION ${HEADER_INSTALL_PATH}/scene
 )
