+++ regen/scene/scene-loader.cpp
@@ -2,12 +2,12 @@
 #include "regen/scene/animation-processor.h"
 #include "regen/scene/shape-processor.h"
 #include "regen/scene/debug-node-processor.h"
-#include "regen/effects/filter.h"
-#include "regen/effects/motion-blur.h"
+#include "regen/passes/filter.h"
+#include "regen/passes/motion-blur.h"
 #include "regen/states/light-pass.h"
 #include "shader-input-processor.h"
-#include "../effects/geometric-picking.h"
-#include "regen/effects/bloom-pass.h"
+#include "../passes/geometric-picking.h"
+#include "regen/passes/bloom-pass.h"
 #include "regen/scene/node-processor.h"
 #include "regen/states/direct-shading.h"
 #include <regen/scene/mesh-processor.h>
