+++ regen/shapes/CMakeLists.txt
@@ -5,5 +5,5 @@ file(GLOB HEADER_FILES *.h)
 
 install(
     FILES ${HEADER_FILES}
-    DESTINATION ${HEADER_INSTALL_PATH}/spatial
+    DESTINATION ${HEADER_INSTALL_PATH}/shapes
 )
